Create README.md
Reads GPS and time data from .gpx files. This works with Garmin and Strave .gpx files, I have not tested on others but it is likely to work. It reads .txt files from yourair (another repository) and combines the files based on timestamps. It displays the results on an interactive map. https://www.garnertotalenergy.com/

It requires Geopanda, they have excellent instructions.

The yourair timestamps must overlay the .gpx timestamps. This part needs work as of this writing.